+++ services/resources/mgmt/2020-05-01/managementgroups/CHANGELOG.md
@@ -1,2 +1,17 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Struct Fields
+
+1. Properties.Path
+
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. Details.Path